+++ dashboard/src/features/orgs/projects/backups/components/BackupListItem/index.ts
@@ -1,2 +1,0 @@
-export * from './BackupListItem';
-export { default as BackupListItem } from './BackupListItem';